third_party/blink/web_tests/svg/custom/touch-events-expected.txt

This test passes if you see a series of PASS lines below:
onclick fired: PASS
ontouchstart fired: PASS
ontouchmove fired: PASS
ontouchend fired: PASS
ontouchcancel fired: PASS
total event count is 6: PASS
